Move search test setup to their own file
